feat(plugin): 完成了核心插件的定义
