+++ packages/plugin/lib/core/env.ts
@@ -1,0 +1,1 @@
+export const pluginName = 'core'
