+++ packages/plugin/lib/core/index.ts
@@ -1,0 +1,19 @@
+import { defineDeltaComicPlugin, useConfig } from '@/index'
+
+import { cfg } from './config'
+export { cfg } from './config'
+
+import { pluginName } from './env'
+import { tokenInit, nativeInit, tokenShare } from './share'
+
+export default defineDeltaComicPlugin(() => ({
+  name: pluginName,
+  i18nName: pluginName,
+  hooks: {
+    onPreboot() {
+      const cs = useConfig()
+      cs.$register(cfg)
+    },
+  },
+  model: { social: { share: { initiative: [tokenInit, nativeInit], tokenListen: [tokenShare] } } },
+}))
