+++ packages/plugin/lib/export/env.ts
@@ -1,0 +1,5 @@
+export type Platform = 'tauri' | 'web'
+export interface ConfigEnv {
+  safe: boolean
+  platform: Platform
+}
