+++ packages/plugin/lib/export/hook/index.ts
@@ -1,4 +1,13 @@
 import type { ContentHooks } from './content'
 export * from './content'
 
-export type PluginConfigHooks = ContentHooks
+import type { RemoteHooks } from './remote'
+export * from './remote'
+
+import type { SocialHooks } from './social'
+export * from './social'
+
+import type { LifecycleHooks } from './lifecycle'
+export * from './lifecycle'
+
+export type PluginConfigHooks = Partial<RemoteHooks & ContentHooks & SocialHooks & LifecycleHooks>
