+++ packages/plugin/lib/export/hook/remote.ts
@@ -1,0 +1,5 @@
+import type { Remote } from '../model'
+
+export interface RemoteHooks {
+  onRemoteTestDone(group: Remote.TestGroup, remote: Remote.Definition | false): void
+}
