+++ packages/plugin/lib/export/hook/social.ts
@@ -1,0 +1,6 @@
+import type { UniItemAuthor } from '@delta-comic/model'
+
+export interface SocialHooks {
+  onSubscribeOne?(author: UniItemAuthor): void
+  onUnsubscribeOne?(author: UniItemAuthor): void
+}
