+++ packages/plugin/lib/export/model/index.ts
@@ -1,8 +1,8 @@
 import type { RemoteModel } from './remote'
 export * as Remote from './remote'
 
-import type { AuthModel } from './auth'
-export * as Auth from './auth'
+import type { UserModel } from './user'
+export * as User from './user'
 
 import type { ExposeModel } from './expose'
 export * as Expose from './expose'
@@ -10,9 +10,17 @@ export * as Expose from './expose'
 import type { ContentModel } from './content'
 export * as Content from './content'
 
+import type { SocialModel } from './social'
+export * as Social from './social'
+
+import type { SpecialModel } from './special'
+export * as Special from './special'
+
 export interface PluginConfigModel {
   remotes?: RemoteModel
-  auth?: AuthModel
-  expose?: ExposeModel
+  user?: UserModel
   content?: ContentModel
+  social?: SocialModel
+  expose?: ExposeModel
+  special?: SpecialModel
 }
