+++ packages/plugin/lib/export/model/special.ts
@@ -1,0 +1,7 @@
+export type SpecialModel = Step[]
+
+export interface Step {
+  name: string
+  async?: boolean
+  call: (setDescription: (description: string) => void) => Promise<void>
+}
