+++ packages/plugin/lib/index.ts
@@ -1,1 +1,5 @@
 export * from './export'
+export * from './module'
+export * from './global'
+
+export * as Core from './core'
