+++ packages/plugin/lib/module/config/index.ts
@@ -1,0 +1,2 @@
+export * from './pointer'
+export * from './store'
