+++ packages/plugin/lib/module/config/pointer.ts
