+++ packages/plugin/lib/module/i18n/index.ts
