+++ pnpm-workspace.yaml
@@ -85,7 +85,7 @@ catalog:
   vite-plugin-wasm: ^3.6.0
   vite-plus: ^0.2.7
   vitest: 5.0.0-beta.7
-  vue: ^3.5.40
+  vue: ^3.6.0-rc.2
   vue-i18n: ^12.0.0-alpha.4
   vue-component-type-helpers: ^3.3.9
   vue-router: ^5.2.0
